requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.35.0
anyio==4.6.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
dataclasses-json==0.6.7
distro==1.9.0
Django==4.2.16
django-cors-headers==4.4.0
djangorestframework==3.15.2
emoji==1.4.2
exceptiongroup==1.2.2
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.9.0
ftfy==6.0.3
google-ai-generativelanguage==0.6.10
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.2
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.2
langchain-community==0.3.1
langchain-core==0.3.9
langchain-text-splitters==0.3.0
langsmith==0.1.131
MarkupSafe==2.1.5
marshmallow==3.22.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.2.1
normalizer==0.0.1
numpy==1.26.4
openai==1.51.0
orjson==3.10.7
packaging==24.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.9.2
pydantic-core==2.23.4
pydantic-settings==2.5.2
pyparsing==3.1.4
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
safetensors==0.4.5
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlparse==0.5.1
sympy==1.13.3
tenacity==8.5.0
tokenizers==0.20.0
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.66.5
transformers==4.45.1
typing-extensions==4.12.2
typing-inspect==0.9.0
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
yarl==1.13.1
